requirements.txt
# core dependencies
python-dotenv
pyyaml

# pdf processing
pdfplumber==0.10.3
PyMuPDF==1.23.8
pdf2image==1.16.3

# ai apis 
openai  
elevenlabs
transformers

# image processing
Pillow
requests

# audio processing
pydub

# video generation
moviepy
numpy

# text-to-speech fallback
gTTS

# stock image APIs
python-unsplash

# cli libraries and utilities
textual>=0.85.0
textual-dev
tqdm

